Moved PaginatedGUI's deprecated methods to the end of the class. Removed 1 unused import. Added abstract Navigable class, which all classes that can have a NavigationBar must extend from, instead of extending directly from GUI. Modified GUIEvents listener to comply with this changes.
